+++ Netable/Netable/LogDestination.swift
@@ -10,6 +10,10 @@ import Foundation
 
 /// Wrapper class for log events emitted by Netable.
 public enum LogEvent: CustomDebugStringConvertible {
+
+    /// Print up some debugging info at start.
+    case startupInfo(baseURL: URL, logDestination: LogDestination)
+
     /// A generic message, not tied to request state.
     case message(StaticString)
 
@@ -25,7 +29,10 @@ public enum LogEvent: CustomDebugStringConvertible {
     /// Default overrides, used by the default logging destination.
     public var debugDescription: String {
         switch self {
-        case .message(let message): return message.description
+        case .startupInfo(let baseURL, let logDestination):
+            return "Netable instance initiated. Here we go! Base URL: \(baseURL.absoluteString). Log Destination: \(logDestination)"
+        case .message(let message):
+            return message.description
         case .requestStarted(let urlString, let method, let headers, let params):
             return "Started \(method.rawValue) request... URL: \(urlString) Headers: \(headers) Params: \(params ?? [:])"
         case .requestCompleted(let statusCode, let responseData, let finalizedResult):